Rename tracking package to tracking_utils and merge facade
- Rename miles/utils/tracking/ to miles/utils/tracking_utils/ to match
  the chat_template_utils/, debug_utils/ naming convention used
  elsewhere under miles/utils/.
- Fold the miles/utils/tracking_utils.py facade (init_tracking, log,
  finish_tracking) into the package's __init__.py so all existing
  callers' import paths
  (`from miles.utils.tracking_utils import init_tracking, ...`) stay
  unchanged.
- Drop the three top-level shim modules
  (miles/utils/{prometheus,tensorboard,wandb}_utils.py); no in-repo
  callers reference them.
