--- miles/utils/tensorboard_utils.py
@@ -1,2 +1,0 @@
-"""Backwards-compatible shim. Prefer ``miles.utils.tracking.tensorboard_utils``."""
-from miles.utils.tracking.tensorboard_utils import _TensorboardAdapter  # noqa: F401
