--- miles/utils/wandb_utils.py
@@ -1,8 +1,0 @@
-"""Backwards-compatible shim. Prefer ``miles.utils.tracking.wandb_utils``."""
-from miles.utils.tracking.wandb_utils import (  # noqa: F401
-    _compute_config_for_logging,
-    _init_wandb_common,
-    _is_offline_mode,
-    init_wandb_primary,
-    init_wandb_secondary,
-)
